use int64_t for block-count deltas instead of ptrdiff_t
`ptrdiff_t` was a side-effect of declaring `GcApplyInfo.block_count_delta`
as Rust `isize` (cheadergen maps `isize` → `ptrdiff_t`). It carries
pointer-arithmetic semantics that aren't appropriate for a block count.
Switch the Rust field to `i64`, the C helper parameter to `int64_t`,
and the four mid-life destruction cast sites from `-(ptrdiff_t)x` to
`-(int64_t)x`. Per reviewer feedback on PR #9735.

+++ src/fork_gc/existing_docs.c
@@ -85,7 +85,7 @@ FGCError FGC_parentHandleExistingDocs(ForkGC *gc) {
     // Sample memory and block count before freeing the index — `InvertedIndex_Free`
     // doesn't know about the owning spec.
     info.bytes_freed += InvertedIndex_MemUsage(idx);
-    IndexStats_BlockCountAdd(&sp->stats, -(ptrdiff_t)InvertedIndex_NumBlocks(idx));
+    IndexStats_BlockCountAdd(&sp->stats, -(int64_t)InvertedIndex_NumBlocks(idx));
     InvertedIndex_Free(idx);
     sp->existingDocs = NULL;
   }
